src/mo2-game-registry-proxy/__init__.py
"""
Copyright (c) Cutleast
"""

import mobase  # pyright: ignore[reportMissingModuleSource]

from .main import GameRegistryProxy


def createPlugin() -> mobase.IPlugin:
    return GameRegistryProxy()
