pyproject.toml
[project]
name = "mo2-game-registry-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyqt6==6.7.1",
]

[dependency-groups]
dev = [
    "mobase-stubs==2.5.2",
    "pyright",
    "ruff",
]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "strict"
exclude = ["**/__pycache__"]
stubPath = "./stubs"
extraPaths = ["./src"]

# Check here for a full list of options: https://github.com/microsoft/pyright/blob/main/docs/configuration.md#diagnostic-settings-defaults
deprecateTypingAliases = false
reportConstantRedefinition = true
reportDeprecated = "warning"
reportDuplicateImport = true
reportIncompleteStub = "warning"
reportIncompatibleMethodOverride = "warning"
reportInconsistentConstructor = "warning"
reportImplicitOverride = "warning"
reportInvalidStubStatement = "warning"
reportMatchNotExhaustive = false
reportMissingImports = "error"
reportMissingParameterType = "error"
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportPrivateUsage = "warning"
reportShadowedImports = "warning"
reportTypeCommentUsage = "information"
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryCast = "information"
reportUnnecessaryComparison = "information"
reportUnnecessaryContains = "information"
reportUnnecessaryIsInstance = "information"
reportUnnecessaryTypeIgnoreComment = "information"
reportUnusedClass = "warning"
reportUnusedImport = "warning"
reportUnusedFunction = "warning"
reportUnusedVariable = "warning"
reportUntypedBaseClass = "error"
reportUntypedClassDecorator = "error"
reportUntypedFunctionDecorator = "error"
reportUntypedNamedTuple = "error"
reportWildcardImportFromLibrary = "warning"
strictDictionaryInference = true
strictListInference = true
strictSetInference = true

